// pacific-sdk/src/code/types.ts
export type VariableValue<T> = {
  type: 'VARIABLE';
  id: string; // E.g. "foo-bar-baz-123"
  fallback: T;
};

export type ConstantValue<T> = {
  type: 'CONSTANT';
  value: T;
};
// ...
export type WorldviewSolidPaint = {
  type: 'SOLID'; // TODO: Opportunity to add other paint types
  color: RGBA;
};

export type WorldviewDropShadowEffect = {
  type: 'DROP_SHADOW';
  color: RGBA;
  offset: Vector;
  radius: number;
  spread?: number; // Why is this undefinable??
};

export type WorldviewInnerShadowEffect = {
  type: 'INNER_SHADOW';
  color: RGBA;
  offset: Vector;
  radius: number;
  spread?: number; // Why is this undefinable??
};

export type WorldviewImagePaint = {
  type: 'IMAGE';
  scaleMode: 'FILL' | 'FIT' | 'CROP' | 'TILE';
  imageHash: string;
};

export type WorldviewGradientPaint = {
  type:
    | 'GRADIENT_ANGULAR'
    | 'GRADIENT_DIAMOND'
    | 'GRADIENT_LINEAR'
    | 'GRADIENT_RADIAL';
  gradientStops: ColorStop[];
  gradientTransform: Transform;
};

export type WorldviewSinglePaint =
  | WorldviewSolidPaint
  | WorldviewImagePaint
  | WorldviewGradientPaint;

export type WorldviewMixedPaint = {
  type: 'MIXED_PAINT';
  values: Array<
    | WorldviewSinglePaint
    | VariableValue<WorldviewSinglePaint>
    | ConstantValue<WorldviewSinglePaint>
  >;
};

export type WorldviewPaint =
  | WorldviewSolidPaint
  | WorldviewImagePaint
  | WorldviewGradientPaint;

export type WorldviewEffect =
  | WorldviewDropShadowEffect
  | WorldviewInnerShadowEffect;
// ...
export type WorldviewContainerNode = {
  type: 'CONTAINER';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;

  //Padding
  paddingTop?: number;
  paddingBottom?: number;
  paddingLeft?: number;
  paddingRight?: number;

  // Layout
  layoutMode?: 'NONE' | 'HORIZONTAL' | 'VERTICAL';
  layoutPositioning?: 'AUTO' | 'ABSOLUTE';
  primaryAxisAlignItems?: 'MIN' | 'MAX' | 'CENTER' | 'SPACE_BETWEEN';
  counterAxisAlignItems: 'MIN' | 'MAX' | 'CENTER' | 'BASELINE';
  itemSpacing?: number;

  // Fill
  fill?:
    | WorldviewPaint
    | VariableValue<WorldviewPaint>
    | ConstantValue<WorldviewPaint>;

  // Stroke
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeTopWeight?: number;
  strokeBottomWeight?: number;
  strokeLeftWeight?: number;
  strokeRightWeight?: number;

  // Corner radius
  topLeftRadius?: number;
  topRightRadius?: number;
  bottomLeftRadius?: number;
  bottomRightRadius?: number;

  // Effects
  effects?: WorldviewEffect[];

  //ProtoType
  reaction?: Reaction[];

  // Children
  children: WorldviewNode[] | null;
};

export type WorldviewComponentNode = {
  type: 'COMPONENT';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;

  // Layout
  layoutMode?: 'NONE' | 'HORIZONTAL' | 'VERTICAL';
  layoutPositioning?: 'AUTO' | 'ABSOLUTE';
  primaryAxisAlignItems?: 'MIN' | 'MAX' | 'CENTER' | 'SPACE_BETWEEN';
  counterAxisAlignItems: 'MIN' | 'MAX' | 'CENTER' | 'BASELINE';

  // Padding
  paddingTop?: number;
  paddingBottom?: number;
  paddingLeft?: number;
  paddingRight?: number;
  itemSpacing?: number;

  // Fill
  fill?:
    | WorldviewPaint
    | VariableValue<WorldviewPaint>
    | ConstantValue<WorldviewPaint>;

  // Stroke
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeTopWeight?: number;
  strokeBottomWeight?: number;
  strokeLeftWeight?: number;
  strokeRightWeight?: number;

  // Corner radius
  topLeftRadius?: number;
  topRightRadius?: number;
  bottomLeftRadius?: number;
  bottomRightRadius?: number;

  // Effects
  effects?: WorldviewEffect[];

  //ProtoType
  reaction?: Reaction[];

  // Children
  children: WorldviewNode[] | null;
};

export type WorldviewInstanceNode = {
  type: 'INSTANCE';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;

  // Layout
  paddingTop?: number;
  paddingBottom?: number;
  paddingLeft?: number;
  paddingRight?: number;

  layoutMode?: 'NONE' | 'HORIZONTAL' | 'VERTICAL';
  layoutPositioning?: 'AUTO' | 'ABSOLUTE';
  primaryAxisAlignItems?: 'MIN' | 'MAX' | 'CENTER' | 'SPACE_BETWEEN';
  counterAxisAlignItems: 'MIN' | 'MAX' | 'CENTER' | 'BASELINE';
  itemSpacing?: number;

  // Fill
  fill?:
    | WorldviewPaint
    | VariableValue<WorldviewPaint>
    | ConstantValue<WorldviewPaint>;

  // Stroke
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeTopWeight?: number;
  strokeBottomWeight?: number;
  strokeLeftWeight?: number;
  strokeRightWeight?: number;

  // Corner radius
  topLeftRadius?: number;
  topRightRadius?: number;
  bottomLeftRadius?: number;
  bottomRightRadius?: number;

  // Effects
  effects?: WorldviewEffect[];

  //ProtoType
  reaction?: Reaction[];

  // Children
  children: WorldviewNode[] | null;
};

export type WorldviewGroupNode = {
  type: 'GROUP';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;

  // Effects
  effects?: WorldviewEffect[];

  //ProtoType
  reaction?: Reaction[];

  // Children
  children: WorldviewNode[] | null;
};

export type WorldviewEllipseNode = {
  type: 'ELLIPSE';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;

  fill?:
    | WorldviewPaint
    | VariableValue<WorldviewPaint>
    | ConstantValue<WorldviewPaint>;

  // Stroke
  // NOTE: This is intended to be recurisve
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeWeight?: number;

  //ProtoType
  reaction?: Reaction[];
};

export type WorldviewLineNode = {
  type: 'LINE';
  id: string;
  name: string;
  x: number;
  y: number;
  rotation: number;
  width: number;
  height: number;
  fill?:
    | WorldviewPaint
    | VariableValue<WorldviewPaint>
    | ConstantValue<WorldviewPaint>;

  // Effects
  effects?: WorldviewEffect[];
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeWeight?: number;

  //ProtoType
  reaction?: Reaction[];
};

export type WorldviewBooleanOperationNode = {
  type: 'BOOLEAN_OPERATION';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;
  fill?:
    | WorldviewPaint
    | VariableValue<WorldviewPaint>
    | ConstantValue<WorldviewPaint>;

  // Effects
  effects?: WorldviewEffect[];

  //ProtoType
  reaction?: Reaction[];

  // Children
  children: WorldviewNode[] | null;
  booleanOperation: 'UNION' | 'SUBTRACT' | 'INTERSECT' | 'EXCLUDE';
};

// TODO: Missing auto layout ...
export type WorldviewVoidContainerNode = {
  type: 'VOID_CONTAINER';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;

  // Fill
  fill?:
    | WorldviewPaint
    | VariableValue<WorldviewPaint>
    | ConstantValue<WorldviewPaint>;

  // Stroke
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeTopWeight?: number;
  strokeBottomWeight?: number;
  strokeLeftWeight?: number;
  strokeRightWeight?: number;

  // Corner radius
  topLeftRadius?: number;
  topRightRadius?: number;
  bottomLeftRadius?: number;
  bottomRightRadius?: number;

  // Effects
  effects?: WorldviewEffect[];

  //ProtoType
  reaction?: Reaction[];
};

export type WorldviewIconContainerNode = {
  type: 'ICON_CONTAINER';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;

  // Fill
  // NOTE: This is intended to be recurisve
  fill?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;

  // Stroke
  // NOTE: This is intended to be recurisve
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeWeight?: number;

  // SVG
  data: string | VariableValue<string> | ConstantValue<string>;

  //ProtoType
  reaction?: Reaction[];
};

export type WorldviewStarNode = {
  type: 'STAR';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;
  rotation: number;
  cornerRadius: number;
  count: number;
  ratio: number;
  cornerSmoothing: number;
  constrainProportions: boolean;
  fill?:
    | WorldviewPaint
    | VariableValue<WorldviewPaint>
    | ConstantValue<WorldviewPaint>;

  // Stroke
  // NOTE: This is intended to be recurisve
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeWeight?: number;

  //ProtoType
  reaction?: Reaction[];
};

export type WorldviewPolygonNode = {
  type: 'POLYGON';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;
  rotation: number;
  cornerRadius: number;
  count: number;
  cornerSmoothing: number;
  constrainProportions: boolean;
  fill?:
    | WorldviewPaint
    | VariableValue<WorldviewPaint>
    | ConstantValue<WorldviewPaint>;

  // Stroke
  // NOTE: This is intended to be recurisve
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeWeight?: number;

  //ProtoType
  reaction?: Reaction[];
};

// TODO: Add textNode.textAutoResize?
// TODO: Add textNode.textDecoration?
export type WorldviewTextNode = {
  type: 'TEXT';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;

  // Fill
  fill?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;

  // Typography
  fontName: FontName; // Sorted
  fontSize: number;
  fontWeight: number;
  letterSpacing: LetterSpacing;
  lineHeight: LineHeight;
  textAlignHorizontal: 'LEFT' | 'CENTER' | 'RIGHT' | 'JUSTIFIED';
  textAlignVertical: 'TOP' | 'CENTER' | 'BOTTOM';
  textCase: 'ORIGINAL' | 'UPPER' | 'LOWER';

  // Characters
  characters: string | VariableValue<string> | ConstantValue<string>;

  // Truncation
  truncation: 'DISABLED' | 'ENDING';

  //ProtoType
  reaction?: Reaction[];
};

export type WorldviewVectorNode = {
  type: 'VECTOR';
  id: string;
  name: string;
  x: number;
  y: number;
  width: number;
  height: number;

  // Fill
  // NOTE: This is intended to be recurisve
  fill?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;

  // Stroke
  // NOTE: This is intended to be recurisve
  stroke?:
    | WorldviewSolidPaint
    | VariableValue<WorldviewSolidPaint>
    | ConstantValue<WorldviewSolidPaint>;
  strokeWeight?: number;

  vectorNetwork: VectorNetwork; // WorldviewVectorNetwork
  vectorPaths: VectorPaths; // Array<WorldviewVectorPath>

  //ProtoType
  reaction?: Reaction[];
};

export type WorldviewNode =
  | WorldviewContainerNode
  | WorldviewComponentNode
  | WorldviewInstanceNode
  | WorldviewVoidContainerNode
  | WorldviewIconContainerNode
  | WorldviewTextNode
  | WorldviewVectorNode
  | WorldviewGroupNode
  | WorldviewEllipseNode
  | WorldviewBooleanOperationNode
  | WorldviewLineNode
  | WorldviewStarNode
  | WorldviewPolygonNode;
// ...
type PreprocessedBinaryAsset = {
  type: 'PREPROCESSED';
  value: WorldviewImagePaint;
};

// Represents a binary asset after processing
type PostprocessedBinaryAsset = {
  type: 'POSTPROCESSED';
  value: {
    url: string;
    description: string;
  };
};

export type EmanationSDKTheme = {
  colors: {
    [key: string]: WorldviewSolidPaint; // TODO?
  };
  images: {
    [key: string]: PreprocessedBinaryAsset | PostprocessedBinaryAsset;
  };
  icons: {
    [key: string]: string;
  };
  characters: {
    [key: string]: string;
  };
  gradients: {
    [key: string]: WorldviewGradientPaint;
  };
};

export type EmanationSDKScreen = {
  meta: {
    theme: EmanationSDKTheme;
  };
  name: string; // E.g. "Profile"
  description: string; // E.g. "The profile screen"
  root: WorldviewNode;
};

export type EmanationSDKApp = {
  meta: {};

  name: string; // E.g. "Finance app"
  description: string; // E.g. "A personal finance app"
  screens: EmanationSDKScreen[];
};

export type EmanationSDKEpoch = {
  name: string; // E.g. "" + Date.now()
  apps: EmanationSDKApp[];
};